requirements.txt
absl-py==2.3.1
Adafruit-Blinka==8.66.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-mlx90640==1.3.6
adafruit-circuitpython-register==1.10.4
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.83.1
Adafruit-PureIO==1.1.11
ArducamDepthCamera==0.1.21
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
astunparse==1.6.3
attrs==25.3.0
av==15.1.0
Babel==2.10.3
beautifulsoup4==4.11.2
binho-host-adapter==0.1.6
blinker==1.5
blobconverter==1.4.3
cachetools==5.5.2
cattrs==25.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.1.0
charset-normalizer==2.0.12
chex==0.1.90
click==8.1.3
colorama==0.4.6
colorzero==2.0
contourpy==1.3.3
coremltools==8.3.0
cryptography==38.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
Deprecated==1.2.18
depthai==2.24.0.0
depthai-sdk==1.9.4
depthai-viewer==0.2.9
dill==0.3.6
distinctipy==1.3.4
distro==1.8.0
docutils==0.19
etils==1.13.0
ffmpy3==0.2.4
filelock==3.19.1
Flask==2.2.2
flatbuffers==20181003210633
flax==0.10.4
fonttools==4.59.2
fsspec==2025.9.0
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
gpiozero==2.0.1
grpcio==1.75.0
h5py==3.14.0
html5lib==1.1
humanize==4.13.0
idna==3.10
importlib-metadata==4.12.0
importlib_resources==6.5.2
isort==5.6.4
itsdangerous==2.1.2
jax==0.4.34
jaxlib==0.4.34
jedi==0.18.2
Jinja2==3.1.2
joblib==1.5.2
jsonpointer==2.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.15.0
keyboard==0.13.5
kiwisolver==1.4.9
lazy-object-proxy==1.9.0
lgpio==0.2.2.0
libarchive-c==5.3
libclang==18.1.1
libevdev==0.5
logilab-common==1.9.8
lxml==4.9.2
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.2
marshmallow==3.17.0
matplotlib==3.10.6
mcap==1.3.0
mcap-ros1-support==0.0.8
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.18
meson==1.5.1
ml-dtypes==0.3.2
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.1
mypy==1.0.1
mypy-extensions==0.4.3
nest_asyncio==1.6.0
networkx==3.5
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
onnx==1.19.0
opencv-contrib-python==4.5.5.62
opencv-python==4.11.0.86
openvino==2025.3.0
openvino-telemetry==2025.2.0
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.5
packaging==23.2
pandas==2.3.2
parso==0.8.3
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.31
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==11.3.0
platformdirs==2.6.0
polars==1.33.1
protobuf==4.25.8
psutil==5.9.4
ptyprocess==0.7.0
pyaml==25.7.0
pyarrow==16.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.23
pycryptodomex==3.11.0
pycups==2.0.1
pyftdi==0.57.1
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.2.4
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-prctl==1.8.1
pytube==15.0.0
PyTurboJPEG==1.6.4
pytz==2022.7.1
pyudev==0.24.0
pyusb==1.2.1
PyYAML==6.0.2
Qt.py==1.4.7
referencing==0.36.2
reportlab==3.6.12
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.18.0
rfc3987==1.3.8
rich==14.1.0
roman==3.3
rosbags==0.9.11
rpds-py==0.27.1
rpi-lgpio==0.6
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
rsa==4.9.1
RTIMULib==7.2.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.2
scipy==1.16.1
Send2Trash==1.8.1b0
sense-hat==2.6.0
sentencepiece==0.2.1
sentry-sdk==1.14.0
simplejpeg==1.8.2
simplejson==3.18.3
six==1.17.0
smbus2==0.4.2
sounddevice==0.5.2
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
sympy==1.14.0
sysv_ipc==1.1.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-cpu-aws==2.15.1
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflowjs==4.22.0
tensorstore==0.1.74
termcolor==3.1.0
tf_keras==2.15.1
thonny==4.1.4
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.11.7
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
treescope==0.1.10
twython==3.8.2
types-aiofiles==22.1
types-annoy==1.17
types-appdirs==1.4
types-aws-xray-sdk==2.10
types-babel==2.11
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.11
types-bleach==5.0
types-boto==2.49
types-braintree==4.17
types-cachetools==5.2
types-caldav==0.10
types-certifi==2021.10.8
types-cffi==1.15
types-chardet==5.0
types-chevron==0.14
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.7
types-contextvars==2.4
types-croniter==1.3
types-cryptography==3.3
types-D3DShot==0.1
types-dateparser==1.1
types-DateTimeRange==1.2
types-decorator==5.1
types-Deprecated==1.2
types-dj-database-url==1.0
types-docopt==0.6
types-docutils==0.19
types-editdistance==0.6
types-emoji==2.1
types-entrypoints==0.4
types-first==2.0
types-flake8-2020==1.7
types-flake8-bugbear==22.10.27
types-flake8-builtins==2.0
types-flake8-docstrings==1.6
types-flake8-plugin-utils==1.3
types-flake8-rst-docstrings==0.2
types-flake8-simplify==0.19
types-flake8-typing-imports==1.14
types-Flask-Cors==3.0
types-Flask-SQLAlchemy==2.5
types-fpdf2==2.5
types-gdb==12.1
types-google-cloud-ndb==1.11
types-hdbcli==2.14
types-html5lib==1.1
types-httplib2==0.21
types-humanfriendly==10.0
types-invoke==1.7
types-JACK-Client==0.5
types-jmespath==1.0
types-jsonschema==4.17
types-keyboard==0.13
types-ldap3==2.9
types-Markdown==3.4
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.1
types-oauthlib==3.2
types-openpyxl==3.0
types-opentracing==2.4
types-paho-mqtt==1.6
types-paramiko==2.11
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.15
types-pep8-naming==0.13
types-Pillow==9.3
types-playsound==1.3
types-polib==1.1
types-prettytable==3.4
types-protobuf==3.20
types-psutil==5.9
types-psycopg2==2.9
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycurl==7.45
types-pyfarmhash==0.3
types-pyflakes==2.5
types-Pygments==2.13
types-pyinstaller==5.6
types-PyMySQL==1.0
types-pynput==1.7
types-pyOpenSSL==22.1
types-pyRFC3339==1.1
types-PyScreeze==0.1
types-pysftp==0.2
types-PySide2==5.15.2.1.7
types-pytest-lazy-fixture==0.6
types-python-crontab==2.6
types-python-dateutil==2.8
types-python-gflags==3.1
types-python-jose==3.3
types-python-nmap==0.7
types-python-slugify==6.1
types-pytz==2022.6
types-pyvmomi==7.0
types-pywin32==304
types-PyYAML==6.0
types-redis==4.3
types-regex==2022.10.31
types-requests==2.28
types-retry==0.9
types-Send2Trash==1.8
types-setuptools==65.5
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-SQLAlchemy==1.4.43
types-stdlib-list==0.8
types-stripe==3.5
types-tabulate==0.9
types-termcolor==1.1
types-toml==0.10
types-toposort==1.7
types-tqdm==4.64
types-tree-sitter==0.20
types-tree-sitter-languages==1.5
types-ttkthemes==3.2
types-typed-ast==1.5
types-tzlocal==4.2
types-ujson==5.5
types-urllib3==1.26
types-vobject==0.9
types-waitress==2.1
types-whatthepatch==1.0
types-xmltodict==0.13
types-xxhash==3.0
types-zxcvbn==4.4
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.202
ultralytics-thop==2.0.17
uritemplate==4.1.1
urllib3==1.26.20
videodev2==0.0.4
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.2
wurlitzer==3.1.1
xmltodict==0.15.0
zipp==1.0.0
zstandard==0.24.0
